% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colfunc.R
\name{colfunc}
\alias{colfunc}
\title{Creates a color scheme for visual field plots}
\usage{
colfunc(n = 35)
}
\arguments{
\item{n}{A positive integer specifying the number of color gradients used in visual field plots and is default to 35}
}
\value{
A color mapping function
}
\description{
\code{colfunc} creates a color scheme visualising the dB values in visual field matrices with darker color corresponding to lower dB values
}
\examples{
colfunc(35)
}
