% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorkey.R
\name{colorkey}
\alias{colorkey}
\title{Generates a color legend for visual field plots}
\usage{
colorkey()
}
\value{
a color legend for dB values in visual field plots
}
\description{
\code{colorkey} generates a color legend for dB values in visual field plots with darker colors corresponding to lower dB values (e.g., black = 0dB, bright yellow >= 35dB)
}
\examples{
colorkey()
}
