% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color.R
\name{get_col}
\alias{get_col}
\title{Color-codes dB values}
\usage{
get_col(db)
}
\arguments{
\item{db}{A number of sensitivity threshold in dB}
}
\value{
the color of a visual field location in hcl color space with darker colors corresponding to lower dB values
}
\description{
\code{get_color} returns the color of a visual field location given its dB value (e.g., black = 0dB, white = 35dB)
}
\examples{
get_col(25)
}
