% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_col.R
\name{get_inv_col}
\alias{get_inv_col}
\title{Color-codes dB values}
\usage{
get_inv_col(db)
}
\arguments{
\item{db}{A number of sensitivity threshold in dB}
}
\value{
either white or black color for a visual field location
}
\description{
\code{get_inv_col} returns the color of a visual field location given its dB value, (e.g., white < 15dB, black > 15dB)
}
\examples{
get_inv_col(25)
}
