% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makevf.R
\name{makevf}
\alias{makevf}
\title{Makes a visual field matrix}
\usage{
makevf(vfvector, eye = c("left", "right"))
}
\arguments{
\item{vfvector}{A vector of length 54. The 54 data points for the 24-2 pattern have to be ordered from superior nasal to inferior temporal.}

\item{eye}{A string of either "left" or "right"}
}
\value{
A matrix of 8 rows and 10 columns
}
\description{
\code{makevf} Makes a visual field matrix from a vector of 54 elements for the 24-2 test results
}
\examples{
vfvector <- rep(35, 54)
makevf(vfvector, 'left')
}
