% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotates a point (x,y) by an angle}
\usage{
rotate(xy, theta = 0)
}
\arguments{
\item{xy}{A vector of length 2 representing the coordinates of a point in cartesian system}

\item{theta}{A number rotating angle in radian}
}
\value{
Coordinates of the point after the rotation
}
\description{
\code{rotate} calculates the coordinates for a point (x,y) after it rotates by an angle theta (radian)
}
\examples{
rotate(c(1000, 0), theta=pi/6)
}
