% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundemo.R
\name{rundemo}
\alias{rundemo}
\title{Runs shiny applications included in the package}
\usage{
rundemo(demo)
}
\arguments{
\item{demo}{A shiny application name. There are two applications i.e., "app", "app2" included in the package}
}
\value{
the called shiny application in a new browser
}
\description{
\code{rundemo(demo)} runs a shiny application provided in the package
}
\examples{
\dontrun{
rundemo("app2.R")
}
}
